+++ pastas/solver/objective/__init__.py
@@ -1,0 +1,7 @@
+# ruff: noqa: F401
+"""Module containing the objective functions used in Pastas for solvers."""
+
+import likelihood
+
+from .likelihood import GaussianLikelihood, GaussianLikelihoodAr1
+from .misfit import misfit
